+++ middleware/isAuthorized.middleware.js
@@ -1,11 +1,55 @@
 const jwt = require('jsonwebtoken');
+const BlogService = require('../service/post.service');
+const { isNumeric } = require('../utils/user.utils');
 
 
-exports.isAuthorized = function(req, res, next) {
+module.exports = {
 
-    const currentUser = req.currentUser;
+    isAuthorized: async(req, res, next) => {
 
-    
+        const currentUserId = req.params.blogId;
+
+        if(!isNumeric(currentUserId)){
+            return res.status(400).json({
+                success: false,
+                message: "Invalid blog id"
+            });
+        }
+
+
+        const result = await BlogService.getBlogByBlogId(currentUserId);
+
+        if(!result){
+            return res.status(404).json({
+                success: false,
+                message: "Blog not found"
+            });
+        }
+
+        const authorOfThisBlog = result.author;
+
+        const accessToken = req.cookies.jwt; 
+        const decodedToken = jwt.verify(accessToken, process.env.ACCESS_TOKEN_SECRET);
+        const currentUser = decodedToken.username;
+
+        try{
+
+            if(!(currentUser==authorOfThisBlog)){
+                return res.status(401).json({
+                    success: false,
+                    message: "Unauthorized"
+                })
+            }
+            
+            next();
+        }
+        catch(e){
+            return res.status(401).send(e);
+        }
+
+    }
+
+
 
 
 }
